Add JS and Wasm syscall handlers
